+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/Filtering/FilterPanel.js
@@ -1,4 +1,10 @@
 /* eslint-disable react/jsx-key */
+/**
+ * Copyright IBM Corp. 2022, 2023
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
 
 import React, { useRef, useMemo, useContext, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
@@ -42,9 +48,9 @@ import { FilterContext } from '.';
 import {
   useInitialStateFromFilters,
   useSubscribeToEventEmitter,
+  useShouldDisableButtons,
 } from './hooks';
 import { getInitialStateFromFilters } from './utils';
-import isEqual from 'lodash/isEqual';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 const componentClass = `${blockClass}-filter-panel`;
@@ -52,7 +58,7 @@ const componentClass = `${blockClass}-filter-panel`;
 const MotionActionSet = motion(ActionSet);
 
 const FilterPanel = ({
-  title,
+  title = 'Filter',
   closeIconDescription = 'Close filter panel',
   updateMethod = BATCH,
   filterSections,
@@ -70,7 +76,6 @@ const FilterPanel = ({
     PANEL
   );
   const [filtersObjectArray, setFiltersObjectArray] = useState([]);
-  const [shouldDisableButtons, setShouldDisableButtons] = useState(true);
   const [showDividerLine, setShowDividerLine] = useState(false);
 
   /** Refs */
@@ -81,6 +86,14 @@ const FilterPanel = ({
   // When using batch actions we have to store the filters to then apply them later
   const prevFiltersRef = useRef(JSON.stringify(filtersState));
   const prevFiltersObjectArrayRef = useRef(JSON.stringify(filtersObjectArray));
+
+  /** State from hooks */
+  const [shouldDisableButtons, setShouldDisableButtons] =
+    useShouldDisableButtons({
+      initialValue: true,
+      filtersState,
+      prevFiltersRef,
+    });
 
   /** Memos */
   const showActionSet = useMemo(() => updateMethod === BATCH, [updateMethod]);
@@ -128,7 +141,11 @@ const FilterPanel = ({
 
   const apply = () => {
     setAllFilters(filtersObjectArray);
+
+    // From the user
     onApply();
+
+    // When the user clicks apply, the action set buttons should be disabled again
     setShouldDisableButtons(true);
 
     // updates the ref so next time the flyout opens we have records of the previous filters
@@ -182,6 +199,19 @@ const FilterPanel = ({
         );
 
         // Remove it from the filters array since there is nothing to filter
+        filtersObjectArrayCopy.splice(index, 1);
+      }
+    } else if (type === DROPDOWN || type === RADIO) {
+      if (value === 'Any') {
+        /**
+        Checks to see if the selected value is 'Any', that means the user wants
+        to reset specific filter
+      */
+        const index = filtersObjectArrayCopy.findIndex(
+          (filter) => filter.id === column
+        );
+
+        // Remove it from the filters array
         filtersObjectArrayCopy.splice(index, 1);
       }
     }
@@ -281,7 +311,11 @@ const FilterPanel = ({
           <FormGroup {...components.FormGroup}>
             <RadioButtonGroup
               {...components.RadioButtonGroup}
-              valueSelected={filtersState[column].value}
+              valueSelected={
+                filtersState[column]?.value === ''
+                  ? 'Any'
+                  : filtersState[column]?.value
+              }
               onChange={(selected) => {
                 setFiltersState({
                   ...filtersState,
@@ -298,6 +332,7 @@ const FilterPanel = ({
                 components.RadioButtonGroup.onChange?.(selected);
               }}
             >
+              <RadioButton id="any" labelText="Any" value="Any" />
               {components.RadioButton.map((radio) => (
                 <RadioButton
                   key={radio.id ?? radio.labelText ?? radio.value}
@@ -311,7 +346,12 @@ const FilterPanel = ({
         return (
           <Dropdown
             {...components.Dropdown}
-            selectedItem={filtersState[column].value}
+            items={['Any', ...components.Dropdown.items]}
+            selectedItem={
+              filtersState[column].value === ''
+                ? 'Any'
+                : filtersState[column].value
+            }
             onChange={({ selectedItem }) => {
               setFiltersState({
                 ...filtersState,
@@ -389,15 +429,6 @@ const FilterPanel = ({
       );
     },
     [filterPanelMinHeight]
-  );
-
-  useEffect(
-    function updateDisabledButtonsState() {
-      setShouldDisableButtons(
-        isEqual(filtersState, JSON.parse(prevFiltersRef.current))
-      );
-    },
-    [filtersState]
   );
 
   useSubscribeToEventEmitter(CLEAR_FILTERS, reset);